Use alter/2 as patch of last resort
If two terms are different one from another but we can find no other
way to represent that difference, describe the difference with
alter/2.  This turned out to simplify the code substantially since it
more closely describes the circumstances in which the old atomic/2
patch should apply.

I changed the name from atomic/2 to alter/2 because it no longer
applies to only atomic terms.  Sometimes it applies to compound terms
which share no structure between them.